#183 add waypoints-state models file
